+++ src/main/java/com/huanqiuyuncang/controller/wms/carton/CartonController.java
@@ -6,6 +6,7 @@ import com.huanqiuyuncang.entity.brand.BrandEntity;
 import com.huanqiuyuncang.entity.carton.CartonEntity;
 import com.huanqiuyuncang.service.wms.carton.CartonInterface;
 import com.huanqiuyuncang.util.AppUtil;
+import com.huanqiuyuncang.util.BeanMapUtil;
 import com.huanqiuyuncang.util.Jurisdiction;
 import com.huanqiuyuncang.util.PageData;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,10 +42,8 @@ public class CartonController extends BaseController {
         String username = Jurisdiction.getUsername();
         Date date = new Date();
         pd.put("cartonid", this.get32UUID());	//主键
-        pd.put("createuser", username);	//创建者
-        pd.put("updateuser", username);	//第一次保存，创建者就是修改者
-        pd.put("createtime", date);	//创建时间
-        pd.put("updatetime", date);	//第一次保存，创建时间就是修改时间
+        BeanMapUtil.setCreateUserInfo(pd);
+        BeanMapUtil.setUpdateUserInfo(pd);
         cartonService.insertSelective(pd);
         mv.addObject("msg","success");
         mv.setViewName("save_result");
@@ -62,8 +61,14 @@ public class CartonController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "del")){return;} //校验权限
         PageData pd = this.getPageData();
         String cartonid = pd.getString("cartonid");
-        cartonService.deleteByPrimaryKey(cartonid);
-        out.write("success");
+        Integer sum = checkTable("wms_carton","cartonID", cartonid);
+        String msg = "success";
+        if(sum >0){
+            msg = "error";
+        }else{
+            cartonService.deleteByPrimaryKey(cartonid);
+        }
+        out.write(msg);
         out.close();
     }
 
@@ -77,6 +82,7 @@ public class CartonController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "edit")){return null;} //校验权限
         ModelAndView mv = this.getModelAndView();
         PageData pd = this.getPageData();
+        BeanMapUtil.setUpdateUserInfo(pd);
         cartonService.updateByPrimaryKeySelective(pd);
         mv.addObject("msg","success");
         mv.setViewName("save_result");
@@ -147,13 +153,16 @@ public class CartonController extends BaseController {
         String DATA_IDS = pd.getString("DATA_IDS");
         if(null != DATA_IDS && !"".equals(DATA_IDS)){
             String ArrayDATA_IDS[] = DATA_IDS.split(",");
+            Integer sum = checkTable("wms_carton","cartonID", ArrayDATA_IDS);
+            if(sum > 0){
+                map.put("msg","error");
+                return AppUtil.returnObject(pd, map);
+            }
             cartonService.deleteAll(ArrayDATA_IDS);
-            pd.put("msg", "ok");
+            map.put("msg", "success");
         }else{
-            pd.put("msg", "no");
+            map.put("msg","error");
         }
-        pdList.add(pd);
-        map.put("list", pdList);
         return AppUtil.returnObject(pd, map);
     }
