+++ src/main/java/com/huanqiuyuncang/controller/wms/customs/CustomsController.java
@@ -62,8 +62,14 @@ public class CustomsController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "del")){return;} //校验权限
         PageData pd = this.getPageData();
         String customsid = pd.getString("customsid");
-        customsService.deleteByPrimaryKey(customsid);
-        out.write("success");
+        Integer sum = checkTable("wms_customs","customsID", customsid);
+        String msg = "success";
+        if(sum >0){
+            msg = "error";
+        }else{
+            customsService.deleteByPrimaryKey(customsid);
+        }
+        out.write(msg);
         out.close();
     }
 
@@ -147,13 +153,16 @@ public class CustomsController extends BaseController {
         String DATA_IDS = pd.getString("DATA_IDS");
         if(null != DATA_IDS && !"".equals(DATA_IDS)){
             String ArrayDATA_IDS[] = DATA_IDS.split(",");
+            Integer sum = checkTable("wms_customs","customsID", ArrayDATA_IDS);
+            if(sum > 0){
+                map.put("msg","error");
+                return AppUtil.returnObject(pd, map);
+            }
             customsService.deleteAll(ArrayDATA_IDS);
-            pd.put("msg", "ok");
+            map.put("msg", "success");
         }else{
-            pd.put("msg", "no");
+            map.put("msg","error");
         }
-        pdList.add(pd);
-        map.put("list", pdList);
         return AppUtil.returnObject(pd, map);
     }
