+++ src/main/java/com/huanqiuyuncang/controller/wms/luggagemail/LuggageMailController.java
@@ -63,8 +63,15 @@ public class LuggageMailController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "del")){return;} //校验权限
         PageData pd = this.getPageData();
         String luggagemailid = pd.getString("luggagemailid");
-        luggageMailService.deleteByPrimaryKey(luggagemailid);
-        out.write("success");
+        Integer sum = checkTable("wms_luggagemail","luggagemailID", luggagemailid);
+        String msg = "success";
+        if(sum >0){
+            msg = "error";
+        }else{
+            luggageMailService.deleteByPrimaryKey(luggagemailid);
+
+        }
+        out.write(msg);
         out.close();
     }
 
@@ -148,13 +155,16 @@ public class LuggageMailController extends BaseController {
         String DATA_IDS = pd.getString("DATA_IDS");
         if(null != DATA_IDS && !"".equals(DATA_IDS)){
             String ArrayDATA_IDS[] = DATA_IDS.split(",");
+            Integer sum = checkTable("wms_luggagemail","luggagemailID", ArrayDATA_IDS);
+            if(sum > 0){
+                map.put("msg","error");
+                return AppUtil.returnObject(pd, map);
+            }
             luggageMailService.deleteAll(ArrayDATA_IDS);
-            pd.put("msg", "ok");
+            map.put("msg", "success");
         }else{
-            pd.put("msg", "no");
+            map.put("msg","error");
         }
-        pdList.add(pd);
-        map.put("list", pdList);
         return AppUtil.returnObject(pd, map);
     }
