+++ src/main/java/com/huanqiuyuncang/controller/wms/order/InnerOrderController.java
@@ -81,8 +81,14 @@ public class InnerOrderController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "del")){return;} //校验权限
         PageData pd = this.getPageData();
         String innerorderid = pd.getString("innerorderid");
-        innerOrderService.deleteByPrimaryKey(innerorderid);
-        out.write("success");
+        Integer sum = checkTable("wms_innerorder","innerorderid", innerorderid);
+        String msg = "success";
+        if(sum >0){
+            msg = "error";
+        }else{
+            innerOrderService.deleteByPrimaryKey(innerorderid);
+        }
+        out.write(msg);
         out.close();
     }
 
@@ -212,13 +218,16 @@ public class InnerOrderController extends BaseController {
         String DATA_IDS = pd.getString("DATA_IDS");
         if(null != DATA_IDS && !"".equals(DATA_IDS)){
             String ArrayDATA_IDS[] = DATA_IDS.split(",");
+            Integer sum = checkTable("wms_innerorder","innerorderid", ArrayDATA_IDS);
+            if(sum > 0){
+                map.put("msg","error");
+                return AppUtil.returnObject(pd, map);
+            }
             innerOrderService.deleteAll(ArrayDATA_IDS);
-            pd.put("msg", "ok");
+            map.put("msg", "success");
         }else{
-            pd.put("msg", "no");
+            map.put("msg","error");
         }
-        pdList.add(pd);
-        map.put("list", pdList);
         return AppUtil.returnObject(pd, map);
     }
