+++ src/main/java/com/huanqiuyuncang/controller/wms/packagetype/PackageTypeController.java
@@ -61,8 +61,14 @@ public class PackageTypeController extends BaseController {
         if(!Jurisdiction.buttonJurisdiction(menuUrl, "del")){return;} //校验权限
         PageData pd = this.getPageData();
         String packageid = pd.getString("packageid");
-        packageTypeService.deleteByPrimaryKey(packageid);
-        out.write("success");
+        Integer sum = checkTable("wms_packagetype","packageId", packageid);
+        String msg = "success";
+        if(sum >0){
+            msg = "error";
+        }else{
+            packageTypeService.deleteByPrimaryKey(packageid);
+        }
+        out.write(msg);
         out.close();
     }
 
@@ -146,13 +152,16 @@ public class PackageTypeController extends BaseController {
         String DATA_IDS = pd.getString("DATA_IDS");
         if(null != DATA_IDS && !"".equals(DATA_IDS)){
             String ArrayDATA_IDS[] = DATA_IDS.split(",");
+            Integer sum = checkTable("wms_packagetype","packageId", ArrayDATA_IDS);
+            if(sum > 0){
+                map.put("msg","error");
+                return AppUtil.returnObject(pd, map);
+            }
             packageTypeService.deleteAll(ArrayDATA_IDS);
-            pd.put("msg", "ok");
+            map.put("msg", "success");
         }else{
-            pd.put("msg", "no");
+            map.put("msg","error");
         }
-        pdList.add(pd);
-        map.put("list", pdList);
         return AppUtil.returnObject(pd, map);
     }
     @RequestMapping(value="/findPackageTypeByPackageType")
